\name{SIMPLE.REGRESSION-package}

\alias{SIMPLE.REGRESSION-package}

\docType{package}

\title{SIMPLE.REGRESSION}

\description{Provides SPSS- and SAS-like output for least squares multiple regression,
    logistic regression, and Poisson regression. Detailed output is also provided for
    OLS moderated regression, interaction plots, and Johnson-Neyman
    regions of significance. The output includes standardized
    coefficients, partial and semi-partial correlations, collinearity diagnostics,
    plots of residuals, and detailed information about simple slopes for interactions. 
    There are numerous options for model plots. 
    \cr\cr The REGIONS_OF_SIGNIFICANCE function also provides
    Johnson-Neyman regions of significance and plots of interactions for both lm
    and lme models (lme models are from the nlme package).
}

\references{
   {Bauer, D. J., & Curran, P. J. (2005). Probing interactions in fixed and multilevel 
	regression: Inferential and graphical techniques. \emph{Multivariate Behavioral 
	Research, 40(3),} 373-400.}
\cr\cr {Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). \emph{Applied 
	multiple regression/correlation analysis for the behavioral sciences} (3rd ed.). 
	Lawrence Erlbaum Associates.}
\cr\cr {Darlington, R. B., & Hayes, A. F. (2017). \emph{Regression analysis and linear models: 
	Concepts, applications, and implementation.} Guilford Press.}
\cr\cr {Dunn, P. K., & Smyth, G. K. (2018). \emph{Generalized linear models 
       with examples in R.} Springer.}
\cr\cr {Hayes, A. F. (2018a). \emph{Introduction to mediation, moderation, and conditional 
	process analysis: A regression-based approach} (2nd ed.). Guilford Press.}
\cr\cr {Huitema, B. (2011). \emph{The analysis of covariance and alternatives: Statistical  
    methods for experiments, quasi-experiments, and single-case studies.} John Wiley & Sons.}
\cr\cr {Johnson, P. O., & Fey, L. C. (1950). The Johnson-Neyman technique, its theory, and 
	application. \emph{Psychometrika, 15,} 349-367.}
\cr\cr {Lorah, J. A. & Wong, Y. J. (2018). Contemporary applications of moderation 
	analysis in counseling psychology. \emph{Counseling Psychology, 65(5),} 629-640.}
\cr\cr {Orme, J. G., & Combs-Orme, T. (2009). \emph{Multiple regression with discrete 
        dependent variables.} Oxford University Press.}
\cr\cr {Pedhazur, E. J. (1997). \emph{Multiple regression in behavioral research: Explanation 
   and prediction.} (3rd ed.). Wadsworth Thomson Learning.}
}
